+++ src/main/java/io/github/mzmine/gui/preferences/MZminePreferences.java
@@ -23,6 +23,7 @@ import io.github.mzmine.main.KeepInMemory;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.parameters.Parameter;
 import io.github.mzmine.parameters.ParameterSet;
+import io.github.mzmine.parameters.dialogs.GroupedParameterSetupDialog;
 import io.github.mzmine.parameters.impl.SimpleParameterSet;
 import io.github.mzmine.parameters.parametertypes.BooleanParameter;
 import io.github.mzmine.parameters.parametertypes.ComboParameter;
@@ -39,6 +40,7 @@ import io.github.mzmine.parameters.parametertypes.submodules.OptionalModuleParam
 import io.github.mzmine.util.ExitCode;
 import java.text.DecimalFormat;
 import java.util.Map;
+import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import org.w3c.dom.Element;
 
@@ -128,35 +130,54 @@ public class MZminePreferences extends SimpleParameterSet {
                                   + "overrides this parameter, if set: --temp D:\\your_tmp_dir\\)",
       System.getProperty("java.io.tmpdir"));
 
-  public static final ComboParameter<KeepInMemory> memoryOption =
-      new ComboParameter<>("Keep in memory",
-          String.format("Specifies the objects that are kept in memory rather than memory mapping "
-                        + "them into temp files in the temp directory. Parameter is overriden by the program "
-                        + "argument --memory. Depending on the read/write speed of the temp directory,"
-                        + " memory mapping is a fast and memory efficient way to handle data, therefore, the "
-                        + "default is to memory map all spectral data and feature data with the option %s. On "
-                        + "systems where memory (RAM) is no concern, viable options are %s and %s, to keep all in memory "
-                        + "or to keep mass lists and feauture data in memory, respectively.",
-              KeepInMemory.NONE, KeepInMemory.ALL, KeepInMemory.MASSES_AND_FEATURES),
-          KeepInMemory.values(), KeepInMemory.NONE);
+  public static final ComboParameter<KeepInMemory> memoryOption = new ComboParameter<>(
+      "Keep in memory", String.format(
+      "Specifies the objects that are kept in memory rather than memory mapping "
+      + "them into temp files in the temp directory. Parameter is overriden by the program "
+      + "argument --memory. Depending on the read/write speed of the temp directory,"
+      + " memory mapping is a fast and memory efficient way to handle data, therefore, the "
+      + "default is to memory map all spectral data and feature data with the option %s. On "
+      + "systems where memory (RAM) is no concern, viable options are %s and %s, to keep all in memory "
+      + "or to keep mass lists and feauture data in memory, respectively.", KeepInMemory.NONE,
+      KeepInMemory.ALL, KeepInMemory.MASSES_AND_FEATURES), KeepInMemory.values(),
+      KeepInMemory.NONE);
 
 
   public MZminePreferences() {
     super(new Parameter[]{
+        // start with performance
+        numOfThreads, memoryOption, tempDirectory, proxySettings, rExecPath, sendStatistics,
+        // visuals
         // number formats
         mzFormat, rtFormat, mobilityFormat, ccsFormat, intensityFormat, ppmFormat, scoreFormat,
         // how to format unit strings
         unitFormat,
         // other preferences
-        numOfThreads, proxySettings, rExecPath, sendStatistics, windowSetttings, sendErrorEMail,
         defaultColorPalette, defaultPaintScale, chartParam, darkMode, presentationMode,
-        imsModuleWarnings, tempDirectory, memoryOption});
+        imsModuleWarnings, windowSetttings, sendErrorEMail});
   }
 
   @Override
   public ExitCode showSetupDialog(boolean valueCheckRequired) {
+    assert Platform.isFxApplicationThread();
+    GroupedParameterSetupDialog dialog = new GroupedParameterSetupDialog(valueCheckRequired, this);
 
-    ExitCode retVal = super.showSetupDialog(valueCheckRequired);
+    // add groups
+    dialog.addParameterGroup("General",
+        new Parameter[]{numOfThreads, memoryOption, tempDirectory, proxySettings, rExecPath,
+            sendStatistics});
+    dialog.addParameterGroup("Formats",
+        new Parameter[]{mzFormat, rtFormat, mobilityFormat, ccsFormat, intensityFormat, ppmFormat,
+            scoreFormat, unitFormat});
+    dialog.addParameterGroup("Visuals",
+        new Parameter[]{defaultColorPalette, defaultPaintScale, chartParam, darkMode,
+            presentationMode});
+    dialog.addParameterGroup("Other",
+        new Parameter[]{imsModuleWarnings, windowSetttings, sendErrorEMail});
+
+    // check
+    dialog.showAndWait();
+    ExitCode retVal = dialog.getExitCode();
 
     if (retVal == ExitCode.OK) {
